requirements.txt
python-dateutil==2.8.2
RandomWords==0.4.0
SQLAlchemy==2.0.28